refactor: tests files
