+++ src/components/error-boundary/error-boundary.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
-import { ErrorBoundary } from '../components/error-boundary/error-boundary';
-import { messages } from '../sources/messages';
+import { ErrorBoundary } from '@components/error-boundary/error-boundary';
+import { messages } from '@sources/messages';
 
 describe('Error boundary', () => {
   it('should catch error and show fallback UI', async () => {
