+++ src/router.ts
@@ -1,7 +1,7 @@
 import { CatDetail } from '@/components/cat-detail/cat-detail';
 import { AboutPage } from '@/pages/about-page/about-page';
 import { HomePage } from '@/pages/home-page/home-page';
-import { NotFoundPage } from '@/pages/not-found-page';
+import { NotFoundPage } from '@/pages/not-found-page/not-found-page';
 import { Layout } from '@components/layout/layout';
 
 import { createBrowserRouter } from 'react-router-dom';
